requirements.txt
google-api-python-client
python-dotenv
mysql-connector-python
pandas
numpy
matplotlib
seaborn
plotly
streamlit
jupyter
notebook
ipykernel
requests
